Fix column type of id.

+++ api/src/main/resources/db/migration/V1.0.0__create_account.sql
@@ -1,7 +1,7 @@
 use simple_journal_entry_db;
 
 create table accounts (
-                       id int not null primary key auto_increment,
+                       id bigint not null primary key auto_increment,
                        code varchar(50) not null unique,
                        name varchar(50) not null,
                        element_type int not null
